Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -3,10 +3,12 @@ package com.educandoweb.course.services;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DatabaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -32,7 +34,18 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		// 1. Verifica se o ID realmente existe no banco
+	    if (!repository.existsById(id)) {
+	        throw new ResourceNotFoundException(id);
+	    }
+	    // 2. Se chegou aqui, existe e tenta deletar
+	    try {
+	        repository.deleteById(id);
+	    } catch (DataIntegrityViolationException e) {
+	        // Esse catch ainda captura se você tentar deletar um usuário 
+	        // que possui pedidos atrelados a ele (erro de chave estrangeira)
+	    	throw new DatabaseException(e.getMessage());
+	    }	
 	}
 	
 	public User update(Long id, User obj) {
